+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/visitor/UnionVisitor.java
@@ -24,14 +24,17 @@ import com.apple.foundationdb.record.RecordCoreException;
 import com.apple.foundationdb.record.RecordMetaData;
 import com.apple.foundationdb.record.metadata.expressions.KeyExpression;
 import com.apple.foundationdb.record.query.expressions.QueryComponent;
+import com.apple.foundationdb.record.query.plan.PlannableIndexTypes;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryFetchFromPartialRecordPlan;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryFilterPlan;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryPlan;
+import com.apple.foundationdb.record.query.plan.plans.RecordQueryPlanWithRequiredFields;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryUnionPlanBase;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 
@@ -59,15 +62,22 @@ import java.util.Set;
  *
  */
 public class UnionVisitor extends RecordQueryPlannerSubstitutionVisitor {
-    public UnionVisitor(@Nonnull final RecordMetaData recordMetadata, @Nullable final KeyExpression commonPrimaryKey) {
-        super(recordMetadata, commonPrimaryKey);
+    public UnionVisitor(@Nonnull final RecordMetaData recordMetadata, @Nonnull final PlannableIndexTypes indexTypes, @Nullable final KeyExpression commonPrimaryKey) {
+        super(recordMetadata, indexTypes, commonPrimaryKey);
     }
 
     @Nonnull
     @Override
-    public RecordQueryPlan postVisit(@Nonnull final RecordQueryPlan recordQueryPlan, @Nonnull Set<KeyExpression> requiredFields) {
+    public RecordQueryPlan postVisit(@Nonnull final RecordQueryPlan recordQueryPlan) {
         if (recordQueryPlan instanceof RecordQueryUnionPlanBase) {
             RecordQueryUnionPlanBase unionPlan = (RecordQueryUnionPlanBase) recordQueryPlan;
+
+            Set<KeyExpression> requiredFields;
+            if (unionPlan instanceof RecordQueryPlanWithRequiredFields) {
+                requiredFields = ((RecordQueryPlanWithRequiredFields)unionPlan).getRequiredFields();
+            } else {
+                requiredFields = Collections.emptySet();
+            }
 
             boolean shouldPullOutFilter = false;
             QueryComponent filter = null;
@@ -93,6 +103,7 @@ public class UnionVisitor extends RecordQueryPlannerSubstitutionVisitor {
                 newChildren.add(newPlan);
             }
             RecordQueryPlan newUnionPlan = new RecordQueryFetchFromPartialRecordPlan(unionPlan.withChildren(newChildren));
+
             if (shouldPullOutFilter) {
                 return new RecordQueryFilterPlan(newUnionPlan, filter);
             } else {